+++ src/components/Fav/FavMobile.js
@@ -20,7 +20,7 @@ import { FixedSizeList as List } from 'react-window';
 import PlaylistAddIcon from '@mui/icons-material/PlaylistAdd';
 
 import { getName } from '@APM/utils/re';
-import useFav from '@hooks/useFav';
+import useFav from './hooks/useFav';
 import { songText } from './FavHeader/FavHeader';
 import { skinPreset } from '../../styles/skin';
 import RandomGIFIcon from './FavHeader/RandomGIF';
