+++ src/components/Fav/SongList/SongList.tsx
@@ -16,8 +16,8 @@ import { zhCN } from '@mui/material/locale';
 import usePlayer from '@hooks/usePlayer';
 import { useNoxSetting } from '@APM/stores/useApp';
 import usePlaylist from '@hooks/usePlaylist';
-import { UseFavP } from '@hooks/useFavPaginated';
-import useRenameSong from '../useRenameSong';
+import { UseFavP } from '../hooks/useFavPaginated';
+import useRenameSong from '../hooks/useRenameSong';
 import { skinPreset } from '../../../styles/skin';
 import SongRenameDialog from '../../dialogs/SongRenameDialog';
 import { ScrollBar } from '../../../styles/styles';
